backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
requests==2.32.3
aiohttp>=3.11.18,<4.0.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.4
python-multipart==0.0.19
google-auth==2.34.0
google-auth-oauthlib==1.2.1
PyJWT==2.10.1
cryptography==44.0.0 
supabase>=2.15.3 